Move "Notify::getTab()" to notification module since it's presentation layer logic
